added some comments to help remember what some of the ajax does.
only display width on image widget blocks since height is variable.

started a detail view for photo that supports json but not using it.  keeping it around for a bit.

added resized images and helper methods for 140 and 600px wide.

cleaned up urls for photos, removed bunch of stuff I don't think we'll ever use. less is better.

adding pixel width choices for select media in page editor so you can choose the right size.  would be nice if it was automatic depending on the widget definition, but for now options are good.